lib/addoninfo.cpp: When loading a JSON addon, test existence and type of script key.
In case a user accidentally uses a wrong JSON file (e.g. naming.json, which is
the config file for namingng.py), the code could give a confusing exception.
This happens when the key 'script' is not defined as a string.

This is solved by testing the key for existence and type. In case 'script' is
not a key or refers to a type other than a string, a clear error is given,
stating for example: 'Loading naming.json failed. script must be set to a
string value.'