addons/namingng.py: Take filename from data.files.
namingng.py attempts to derive the source filename from the name of the
dumpfile. However, the dumpfile is not necessarily named according to this
pattern, e.g. cppcheck will add the pid to the filename. Taking the first item
of data.files seem to be more robust.

To get the basename of the file, os.path.basename() is used. This solves
(theoretical) issues on platforms with a different path separator.