addons/namingng.py: Improve file name checking feature.
With this patch, all filenames are checked, not just those provided on the
cppcheck command line. This is useful as header files will now also be part of
this check.

The "RE_FILE" key of the configuration JSON may contain a list of regular
expressions, where any match will lead to acceptance of the filename.

Both the full path and the basename of the files are tested.

One use case for this combination of features is:

"RE_FILE":[
    "/.*\\.h\\Z",
    "[a-z][a-z0-9_]*[a-z0-9]\\.[ch]\\Z"
]

This will accept any file naming convention of the platform used (assuming
platform files are all referenced using an absolute path), while enforcing
a particular naming scheme for project files.